Make verify_used_files work with stdlib.
